Made killing of subprocesses configurable
Before this commit, signals were sent to all children of a process.
This may lead to undesired effects. For example, if an application was
started through a terminal and the terminal will be stopped, the
application started through that terminal will also be stopped.

Subprocess that should also be killed must be explicitly configured now.